More test fixes
and splitting up large server_agent file
