+++ tests/agent_server/test_server_agent.py
@@ -278,7 +278,7 @@ async def test_deploy_empty(server, client, resource_container, environment):
 
 
 @pytest.mark.asyncio(timeout=100)
-async def test_deploy_with_undefined(server_multi, client_multi, resource_container):
+async def test_deploy_with_undefined(server_multi, client_multi, resource_container, async_finalizer):
     """
          Test deploy of resource with undefined
     """
@@ -306,6 +306,7 @@ async def test_deploy_with_undefined(server_multi, client_multi, resource_contai
         agent_map={"agent1": "localhost", "agent2": "localhost"},
         code_loader=False,
     )
+    async_finalizer.add(agent.stop)
     agent.add_end_point_name("agent2")
     await agent.start()
 
@@ -415,8 +416,6 @@ async def test_deploy_with_undefined(server_multi, client_multi, resource_contai
         )
 
     await retry_limited(done, 100)
-
-    await agent.stop()
     inmanta.agent.agent.GET_RESOURCE_BACKOFF = backoff
 
 
@@ -2171,350 +2170,6 @@ async def test_dryrun_failures(resource_container, server, client):
     )
 
     await agent.stop()
-
-
-@pytest.mark.asyncio
-async def test_send_events(resource_container, environment, server, client):
-    """
-        Send and receive events within one agent
-    """
-    agentmanager = server.get_slice(SLICE_AGENT_MANAGER)
-
-    resource_container.Provider.reset()
-    agent = Agent(
-        hostname="node1",
-        environment=environment,
-        agent_map={"agent1": "localhost"},
-        code_loader=False,
-    )
-    agent.add_end_point_name("agent1")
-    await agent.start()
-    await retry_limited(lambda: len(agentmanager.sessions) == 1, 10)
-
-    version = int(time.time())
-
-    res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
-    resources = [
-        {
-            "key": "key1",
-            "value": "value1",
-            "id": res_id_1,
-            "send_event": False,
-            "purged": False,
-            "requires": ["test::Resource[agent1,key=key2],v=%d" % version],
-        },
-        {
-            "key": "key2",
-            "value": "value2",
-            "id": "test::Resource[agent1,key=key2],v=%d" % version,
-            "send_event": True,
-            "requires": [],
-            "purged": False,
-        },
-    ]
-
-    result = await client.put_version(
-        tid=environment,
-        version=version,
-        resources=resources,
-        unknowns=[],
-        version_info={},
-    )
-    assert result.code == 200
-
-    # do a deploy
-    result = await client.release_version(
-        environment, version, True, const.AgentTriggerMethod.push_full_deploy
-    )
-    assert result.code == 200
-
-    result = await client.get_version(environment, version)
-    assert result.code == 200
-
-    await _wait_until_deployment_finishes(client, environment, version)
-
-    events = resource_container.Provider.getevents("agent1", "key1")
-    assert len(events) == 1
-    for res_id, res in events[0].items():
-        assert res_id.agent_name == "agent1"
-        assert res_id.attribute_value == "key2"
-        assert res["status"] == const.ResourceState.deployed
-        assert res["change"] == const.Change.created
-
-    await agent.stop()
-
-
-@pytest.mark.asyncio
-async def test_send_events_cross_agent(resource_container, environment, server, client):
-    """
-        Send and receive events over agents
-    """
-    agentmanager = server.get_slice(SLICE_AGENT_MANAGER)
-
-    resource_container.Provider.reset()
-    agent = Agent(
-        hostname="node1",
-        environment=environment,
-        agent_map={"agent1": "localhost"},
-        code_loader=False,
-    )
-    agent.add_end_point_name("agent1")
-    await agent.start()
-    await retry_limited(lambda: len(agentmanager.sessions) == 1, 10)
-
-    agent2 = Agent(
-        hostname="node2",
-        environment=environment,
-        agent_map={"agent2": "localhost"},
-        code_loader=False,
-    )
-    agent2.add_end_point_name("agent2")
-    await agent2.start()
-    await retry_limited(lambda: len(agentmanager.sessions) == 2, 10)
-
-    version = int(time.time())
-
-    res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
-    resources = [
-        {
-            "key": "key1",
-            "value": "value1",
-            "id": res_id_1,
-            "send_event": False,
-            "purged": False,
-            "requires": ["test::Resource[agent2,key=key2],v=%d" % version],
-        },
-        {
-            "key": "key2",
-            "value": "value2",
-            "id": "test::Resource[agent2,key=key2],v=%d" % version,
-            "send_event": True,
-            "requires": [],
-            "purged": False,
-        },
-    ]
-
-    result = await client.put_version(
-        tid=environment,
-        version=version,
-        resources=resources,
-        unknowns=[],
-        version_info={},
-    )
-    assert result.code == 200
-
-    # do a deploy
-    result = await client.release_version(
-        environment, version, True, const.AgentTriggerMethod.push_full_deploy
-    )
-    assert result.code == 200
-
-    result = await client.get_version(environment, version)
-    assert result.code == 200
-
-    await _wait_until_deployment_finishes(client, environment, version)
-
-    assert resource_container.Provider.get("agent1", "key1") == "value1"
-    assert resource_container.Provider.get("agent2", "key2") == "value2"
-
-    events = resource_container.Provider.getevents("agent1", "key1")
-    assert len(events) == 1
-    for res_id, res in events[0].items():
-        assert res_id.agent_name == "agent2"
-        assert res_id.attribute_value == "key2"
-        assert res["status"] == const.ResourceState.deployed
-        assert res["change"] == const.Change.created
-
-    await agent.stop()
-    await agent2.stop()
-
-
-@pytest.mark.asyncio
-async def test_send_events_cross_agent_deploying(
-    resource_container, environment, server, client, no_agent_backoff
-):
-    """
-        Send and receive events over agents
-    """
-    agentmanager = server.get_slice(SLICE_AGENT_MANAGER)
-
-    resource_container.Provider.reset()
-    agent = Agent(
-        hostname="node1",
-        environment=environment,
-        agent_map={"agent1": "localhost"},
-        code_loader=False,
-    )
-    agent.add_end_point_name("agent1")
-    await agent.start()
-    await retry_limited(lambda: len(agentmanager.sessions) == 1, 10)
-
-    agent2 = Agent(
-        hostname="node2",
-        environment=environment,
-        agent_map={"agent2": "localhost"},
-        code_loader=False,
-    )
-    agent2.add_end_point_name("agent2")
-    await agent2.start()
-    await retry_limited(lambda: len(agentmanager.sessions) == 2, 10)
-
-    version = int(time.time())
-
-    res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
-    resources = [
-        {
-            "key": "key1",
-            "value": "value1",
-            "id": res_id_1,
-            "send_event": False,
-            "purged": False,
-            "requires": ["test::Wait[agent2,key=key2],v=%d" % version],
-        },
-        {
-            "key": "key2",
-            "value": "value2",
-            "id": "test::Wait[agent2,key=key2],v=%d" % version,
-            "send_event": True,
-            "requires": [],
-            "purged": False,
-        },
-    ]
-
-    result = await client.put_version(
-        tid=environment,
-        version=version,
-        resources=resources,
-        unknowns=[],
-        version_info={},
-    )
-    assert result.code == 200
-
-    # do a deploy
-    result = await client.release_version(
-        environment, version, True, const.AgentTriggerMethod.push_full_deploy
-    )
-    assert result.code == 200
-
-    result = await client.get_version(environment, version)
-    assert result.code == 200
-
-    await _wait_for_n_deploying(client, environment, version, 1)
-
-    # restart deploy
-    result = await client.release_version(
-        environment, version, True, const.AgentTriggerMethod.push_full_deploy
-    )
-    assert result.code == 200
-
-    await resource_container.wait_for_done_with_waiters(client, environment, version)
-
-    # incorrect CAD handling causes skip, which completes deploy without writing
-    assert resource_container.Provider.get("agent1", "key1") == "value1"
-
-    await agent.stop()
-    await agent2.stop()
-
-
-@pytest.mark.asyncio(timeout=15)
-async def test_send_events_cross_agent_restart(
-    resource_container, environment, server, client, no_agent_backoff
-):
-    """
-        Send and receive events over agents with agents starting after deploy
-    """
-    agentmanager = server.get_slice(SLICE_AGENT_MANAGER)
-
-    config.Config.set("config", "agent-deploy-interval", "0")
-    config.Config.set("config", "agent-repair-interval", "0")
-    resource_container.Provider.reset()
-    agent2 = Agent(
-        hostname="node2",
-        environment=environment,
-        agent_map={"agent2": "localhost"},
-        code_loader=False,
-    )
-    agent2.add_end_point_name("agent2")
-    await agent2.start()
-    await retry_limited(lambda: len(agentmanager.sessions) == 1, 10)
-
-    version = int(time.time())
-
-    res_id_1 = "test::Resource[agent1,key=key1],v=%d" % version
-    resources = [
-        {
-            "key": "key1",
-            "value": "value1",
-            "id": res_id_1,
-            "send_event": False,
-            "purged": False,
-            "requires": ["test::Resource[agent2,key=key2],v=%d" % version],
-        },
-        {
-            "key": "key2",
-            "value": "value2",
-            "id": "test::Resource[agent2,key=key2],v=%d" % version,
-            "send_event": True,
-            "requires": [],
-            "purged": False,
-        },
-    ]
-
-    result = await client.put_version(
-        tid=environment,
-        version=version,
-        resources=resources,
-        unknowns=[],
-        version_info={},
-    )
-    assert result.code == 200
-
-    # do a deploy
-    result = await client.release_version(
-        environment, version, True, const.AgentTriggerMethod.push_full_deploy
-    )
-    assert result.code == 200
-
-    result = await client.get_version(environment, version)
-    assert result.code == 200
-
-    # wait for agent 2 to finish
-    while (result.result["model"]["total"] - result.result["model"]["done"]) > 1:
-        result = await client.get_version(environment, version)
-        await asyncio.sleep(1)
-
-    assert resource_container.Provider.get("agent2", "key2") == "value2"
-
-    # start agent 1 and wait for it to finish
-    agent = Agent(
-        hostname="node1",
-        environment=environment,
-        agent_map={"agent1": "localhost"},
-        code_loader=False,
-    )
-    agent.add_end_point_name("agent1")
-    await agent.start()
-    await retry_limited(lambda: len(agentmanager.sessions) == 2, 10)
-
-    # Events are only propagated in a full deploy
-    await agent._instances["agent1"].get_latest_version_for_agent(
-        reason="Repair", incremental_deploy=False, is_repair_run=False
-    )
-
-    await _wait_until_deployment_finishes(client, environment, version)
-
-    assert resource_container.Provider.get("agent1", "key1") == "value1"
-
-    events = resource_container.Provider.getevents("agent1", "key1")
-    assert len(events) == 1
-    for res_id, res in events[0].items():
-        assert res_id.agent_name == "agent2"
-        assert res_id.attribute_value == "key2"
-        assert res["status"] == const.ResourceState.deployed
-        assert res["change"] == const.Change.created
-
-    await agent.stop()
-    await agent2.stop()
 
 
 @pytest.mark.parametrize(
